+++ src/Sockets/UnknownJsonRpcRequestMethodException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Serenata\Sockets;
+
+use Exception;
+
+/**
+ * Indicates the specified JSON-RPC request method is not known.
+ */
+final class UnknownJsonRpcRequestMethodException extends Exception
+{
+}